model/requirements.txt
accelerate=0.18.0
dill==0.3.4
dominate==2.7.0
easydict==1.9
Flask==2.1.0
Flask_Cors==3.0.9
gym==0.26.2
matplotlib==3.3.2
numpy==1.22.0
onnx==1.13.0
onnx_caffe2==1.0.0
opencv_python==4.5.3.56
Pillow==9.4.0
pycocotools==2.0.6
python-dotenv==0.21.1
requests==2.28.1
scipy==1.7.0
six==1.16.0
skimage==0.0
tensorflow==2.11.0
torch==1.9.0
torchtext==0.14.1
torchvision==0.10.0
